Refactor the code + Update the workflow
