+++ app/repositories/interfaces/food.go
@@ -1,0 +1,9 @@
+package interfaces
+
+import (
+	"foodapplication/models"
+)
+
+type FoodRepository interface {
+	CreateFood(food *models.Food) (int, error)
+}
