--- app/repositories/interfaces/ifoodrepository.go
@@ -1,9 +1,0 @@
-package interfaces
-
-import (
-	"foodapplication/models"
-)
-
-type IFoodRepository interface {
-	CreateFood(food *models.Food) (int, error)
-}
