+++ app/services/interfaces/food.go
@@ -1,0 +1,7 @@
+package interfaces
+
+import "foodapplication/models"
+
+type FoodService interface {
+	CreateFood(food *models.Food) (int, error)
+}
