--- app/services/interfaces/ifoodservice.go
@@ -1,7 +1,0 @@
-package interfaces
-
-import "foodapplication/models"
-
-type IFoodService interface {
-	CreateFood(food *models.Food) (int, error)
-}
